refactoring Render file
